Remove support for .NET 2.0 and 3.0
For a source code-only distribution, re-defining the `[Extension]`
attribute for target frameworks that do not have it (.NET <3.5) is
potentially going to cause collisions if the target project already
does this.

A fairly clean option would be to have the source NuGet package de-
pend on another source code-only package that provides this attribute.
But for now, we simply remove it altogether since probably noone uses
.NET 3.0 and earlier anymore anyways.