Add <TypeNameFormatterInternal> MSBuild property
For the source code distribution, the library's types should be de-
clared `internal` so they don't spill out of the referencing project.
(This is unlike with a binary distribution, where they would have to
be declared `public` to be visible outside their own assembly.)

Introduce a MSBuild property <TypeNameFormatterInternal> that defaults
to "True" in the source code NuGet package.