Prevent debugger from "stepping into" (F11)
Applying [DebuggerStepThrough] to the class means that the debugger
won't "step into" (F11) TypeNameFormatter (even though it will still
honour breakpoints set in there), so this makes the library less of
an obstacle while still allowing it to be debugged.

Another option considered was [DebuggerNonUserCode] which works hand
in hand with the "Enable Just My Code" debugger setting, and while it
would be appropriate (given the name), it is a very coarse-grained
approach. If someone wants to debug TypeNameFormatter, they shouldn't
have to disable JMC and possibly step into a whole lot of other non-
user code.

[DebuggerStepThrough] has the better tradeoff here (IMHO).