Disallow omission of type params where used as arg
This requires a slightly larger refactoring. For nested types, it is
no longer sufficient to forward an array with the generic type params/
args, as this doesn't allow us to find out whether generic "params"
are to be interpreted as parameters or arguments. We can only figure
that out if we know whether the "innermost", enclosed type is a con-
structed generic type or a type definition... so instead of forwarding
generic params, we forward the innermost type instead.

This also restructures the main formatting method so the two phases
(filtering out special/composite cases, actual formatting) are more
easily distinguishable.