Invert definiton of `GenericParameterNames` flag
This turns the flag into `NoGenericParameterNames`. That is, generic
parameter names are now included by default and one has to explicitly
opt out to have them omitted.

The reason for this is to make it less likely for people to stumble on
output that will seem like a bug in the library, but is really caused
by a limitation of Reflection. There are some cases where we cannot
determine whether we are dealing with a generic type parameter (which
depending on the flags you might want to omit), or with a type argument
(which should never be omitted). Reflection will incorrectly tend
towards the former:

    class Type<T> {
        Type<T> Field;
    }

Formatting the type's name can be `Type<T>` or `Type<>` depending on
the options, but formatting the field's type should always yield
`Type<T>` because it's actually a "semi"-constructed generic type
instantiated over its own type argument. But we cannot detect this.