Rewrite .targets to better support non-VS IDEs
It was reported that TypeNameFormatter doesn't work well in Rider.
This appears to be due to the fact that the source files are only
registered for compilation (via <Compile Include="..."/>) inside a
custom build target, and Rider's IDE won't detect those properly.

Let's rewrite the .targets file so that the <Compile Include> happen
outside any build target. This has the somewhat unfortunate effect
that a folder node will become visible in VS' Solution Explorer that
cannot be fully hidden.